Move Gradle Home and Gradle Profiler directories up.
We've noticed a 5-10% performance improvement in certain scenarios because of the way file system watching will watch the entire directory. It is also not a typical setup that the gradle-user-home is also in the project directory.

Because of this, the `output` directory in the root folder has been renamed to `results` to separate build outputs from results.